Improved "Select Vertices by Connections" node
